created the state.js file that holds the State class.  The state class is our gamestate.  I also created the routes/game.js file which will server as our API endpoints. Anything related to the games logic will go inside this file.  There should only be a handful of endpoints needed to make the game function.
