+++ app.js
@@ -5,11 +5,13 @@ const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
 // added modules
-const States = require('./sequelize');
+const sequelize = require('./sequelize');
+const GameState = require("./state.js");
 const uuidv4 = require('uuid/v4');
 
 const indexRouter = require('./routes/index');
 const usersRouter = require('./routes/users');
+const gameRouter = require('./routes/game')
 
 const app = express();
 
@@ -25,6 +27,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
+app.use('/game', gameRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -42,35 +45,18 @@ app.use(function(err, req, res, next) {
   res.render('error');
 });
 
-//TESTING OUT UUID, and States objects to put a json object in the database
-initGameState = (_uuid) => {
-  var o = {} // empty Object
-  var key = 'state';
-  o[key] = []; // empty Array, which you can push() values into
-  var data = {
-      uuid:_uuid,
-      playerCount:0,
-      pot: 0,
-      deck:[],
-      players:[],
-      currentPlayer:'none'
-  };
-  o[key].push(data);
-  return JSON.stringify(o);
-}
 
-let id = uuidv4();
-let json = initGameState(id);
 
-console.log(json);
+// let tempState = new GameState(uuidv4());
 
-// using States.create is one way of inputing data into the database.
-// we will use pg-promise for other types of queries
-States.create({
-  uuid: id,
-  state: json
-})
-  .then(() => console.log("new entry has been created in db"));
-//TESTING OUT UUID, and States objects to put a json object in the database
+// // using States.create is one way of inputing data into the database.
+// // we will use pg-promise for other types of queries
+// sequelize.create({
+//   uuid: tempState.uuid,
+//   state: tempState.json
+// })
+//   .then(() => console.log("new entry has been created in db"));
+// //TESTING OUT UUID, and States objects to put a json object in the database
+
 
 module.exports = app;
